use longs instead of ints to store pointers in NIO buffers
This allows the code to be used on 64-bit VMs.

+++ luni/src/main/java/java/nio/DirectByteBuffer.java
@@ -33,7 +33,7 @@ abstract class DirectByteBuffer extends BaseByteBuffer {
         }
 
         this.offset = offset;
-        this.effectiveDirectAddress = block.toInt() + offset;
+        this.effectiveDirectAddress = block.toLong() + offset;
     }
 
     @Override
